+++ GANDLF/compute/forward_pass.py
@@ -3,14 +3,15 @@ import torch
 from tqdm import tqdm
 import SimpleITK as sitk
 import numpy as np
+import pandas as pd
 import torchio
 
 from GANDLF.utils import (
     get_date_time,
     get_filename_extension_sanitized,
     reverse_one_hot,
-    resample_image,
 )
+from GANDLF.post_process import fill_holes
 from .step import step
 from .loss_and_metric import get_loss_and_metrics
 
@@ -55,6 +56,7 @@ def validate_network(
             total_epoch_valid_metric[metric] = 0
 
     logits_list = []
+    subject_id_list = []
     is_classification = params.get("problem_type") == "classification"
     is_inference = mode == "inference"
 
@@ -118,6 +120,7 @@ def validate_network(
                     path=subject["label"]["path"],
                     type=torchio.LABEL,
                     tensor=subject["label"]["data"].squeeze(0),
+                    affine=subject["label"]["affine"].squeeze(0),
                 )
                 label_present = True
                 label_ground_truth = subject_dict["label"]["data"]
@@ -134,6 +137,7 @@ def validate_network(
                 path=subject[key]["path"],
                 type=subject[key]["type"],
                 tensor=subject[key]["data"].squeeze(0),
+                affine=subject[key]["affine"].squeeze(0),
             )
 
         # regression/classification problem AND label is present
@@ -163,6 +167,7 @@ def validate_network(
 
             if is_inference and is_classification:
                 logits_list.append(pred_output)
+                subject_id_list.append(subject.get("subject_id")[0])
 
             if params["save_output"]:
                 outputToWrite += (
@@ -265,31 +270,42 @@ def validate_network(
                 output_prediction = output_prediction.unsqueeze(0)
                 label_ground_truth = label_ground_truth.unsqueeze(0).to(torch.float32)
                 if params["save_output"]:
-                    path_to_metadata = subject["path_to_metadata"][0]
-                    inputImage = sitk.ReadImage(path_to_metadata)
-                    ext = get_filename_extension_sanitized(path_to_metadata)
+                    img_for_metadata = torchio.Image(
+                        type=subject["label"]["type"],
+                        tensor=subject["label"]["data"].squeeze(0),
+                        affine=subject["label"]["affine"].squeeze(0),
+                    ).as_sitk()
+                    ext = get_filename_extension_sanitized(
+                        subject["path_to_metadata"][0]
+                    )
                     pred_mask = output_prediction.numpy()
                     # '0' because validation/testing dataloader always has batch size of '1'
                     pred_mask = reverse_one_hot(
                         pred_mask[0], params["model"]["class_list"]
                     )
                     pred_mask = np.swapaxes(pred_mask, 0, 2)
+                    # perform numpy-specific postprocessing here
+                    if "fill_holes" in params["data_postprocessing"]:
+                        pred_mask = fill_holes(pred_mask)
+
                     ## special case for 2D
                     if image.shape[-1] > 1:
-                        # ITK expects array as Z,X,Y
                         result_image = sitk.GetImageFromArray(pred_mask)
                     else:
                         result_image = sitk.GetImageFromArray(pred_mask.squeeze(0))
-                    result_image.CopyInformation(inputImage)
+                    result_image.CopyInformation(img_for_metadata)
+
                     # cast as the same data type
-                    result_image = sitk.Cast(result_image, inputImage.GetPixelID())
+                    result_image = sitk.Cast(
+                        result_image, img_for_metadata.GetPixelID()
+                    )
                     # this handles cases that need resampling/resizing
                     if "resample" in params["data_preprocessing"]:
-                        result_image = resample_image(
-                            result_image,
-                            inputImage.GetSpacing(),
-                            interpolator=sitk.sitkNearestNeighbor,
+                        resampler = torchio.transforms.Resample(
+                            img_for_metadata.GetSpacing(),
+                            interpolator=sitk.NearestNeighbor,
                         )
+                        result_image = resampler(result_image)
                     sitk.WriteImage(
                         result_image,
                         os.path.join(
@@ -320,6 +336,7 @@ def validate_network(
             output_prediction = output_prediction.squeeze(-1)
             if is_inference and is_classification:
                 logits_list.append(output_prediction)
+                subject_id_list.append(subject.get("subject_id")[0])
 
             # we cast to float32 because float16 was causing nan
             final_loss, final_metric = get_loss_and_metrics(
@@ -399,12 +416,17 @@ def validate_network(
     # write the predictions, if appropriate
     if params["save_output"]:
         if is_inference and is_classification and logits_list:
+            class_list = [str(c) for c in params["model"]["class_list"]]
             logit_tensor = torch.cat(logits_list)
             current_fold_dir = params["current_fold_dir"]
-            np.savetxt(
-                os.path.join(current_fold_dir, "logits.csv"),
-                logit_tensor.detach().cpu().numpy(),
-                delimiter=",",
+            logit_tensor = logit_tensor.detach().cpu().numpy()
+            columns = ["SubjectID"] + class_list
+            logits_df = pd.DataFrame(columns=columns)
+            logits_df.SubjectID = subject_id_list
+            logits_df[class_list] = logit_tensor
+
+            logits_df.to_csv(
+                os.path.join(current_fold_dir, "logits.csv"), index=False, sep=","
             )
 
         if "value_keys" in params: