+++ GANDLF/metrics/__init__.py
@@ -10,9 +10,10 @@ from .segmentation import (
     hd95,
     hd95_per_label,
 )
-from .regression import accuracy, classification_accuracy, balanced_acc_score
-from .generic import recall_score, precision_score, iou_score, f1_score
 from .GAN import FID_score, SSIM_score, LPIPS_score, NMI_score
+from .regression import classification_accuracy, balanced_acc_score
+from .generic import recall_score, precision_score, iou_score, f1_score, accuracy
+
 
 # global defines for the metrics
 
@@ -23,7 +24,7 @@ global_metrics_dict = {
     "mse": MSE_loss,
     "hd95": hd95,
     "hd95_per_label": hd95_per_label,
-    "hausdorff95": hd100,
+    "hausdorff95": hd95,
     "hd100": hd100,
     "hd100_per_label": hd100_per_label,
     "hausdorff": hd100,