+++ GANDLF/models/__init__.py
@@ -5,6 +5,7 @@ from .uinc import uinc
 from .fcn import fcn
 from .vgg import vgg11, vgg13, vgg16, vgg19
 from .densenet import densenet121, densenet169, densenet201, densenet264
+from .resnet import resnet18, resnet34, resnet50, resnet101, resnet152
 from .sdnet import SDNet
 from .MSDNet import MSDNet
 from .brain_age import brainage
@@ -40,6 +41,11 @@ global_models_dict = {
     "densenet264": densenet264,
     "msdnet": MSDNet,
     "brain_age": brainage,
+    "resnet18": resnet18,
+    "resnet34": resnet34,
+    "resnet50": resnet50,
+    "resnet101": resnet101,
+    "resnet152": resnet152,
 }
 
 global_gan_models_dict = {
@@ -48,4 +54,5 @@ global_gan_models_dict = {
     "pix2pixHD": pix2pixHD,
     "cycleGAN": cycleGAN,
     "dcgan": dcgan,
+    "sdnet": SDNet,
 }